+++ intercepting-filter/src/main/Filter.java
@@ -1,0 +1,8 @@
+/**
+ * Filter interface
+ * @author joshzambales
+ *
+ */
+public interface Filter{
+	public String execute(String[] request);
+}
