+++ intercepting-filter/src/main/OrderFilter.java
@@ -1,0 +1,13 @@
+/**
+ * Concrete implementation of filter
+ * 
+ * @author joshzambales
+ *
+ */
+public class OrderFilter implements Filter{
+	public String execute(String[] request){
+		if(request[4].equals("")){
+			return null;
+		}else return request[4];
+	}
+}
