+++ pom.xml
@@ -41,6 +41,7 @@
 		<module>null-object</module>
 		<module>event-aggregator</module>
 		<module>callback</module>
+		<module>intercepting-filter</module>
   </modules>
 
 	<dependencyManagement>
